Created main application layout.
