+++ ScoutSuite/frontend/src/layout/Header/style.scss
@@ -1,0 +1,27 @@
+@import 'src/variables';
+
+.nav-header {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  background-color: $ncc-grey-dark;
+
+  .content {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 2rem;
+    margin: auto  0;
+
+    img {
+      height: 2.75rem;
+    }
+  }
+
+  &::after {
+    content: ' ';
+    height: 0.625rem;
+    width: 100%;
+    background-color: $ncc-red;
+  }
+}
