+++ ScoutSuite/frontend/src/layout/Menu/index.js
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import './style.scss';
+
+const Menu = () => {
+  return (
+    <div className="nav-menu">
+      <span>Fake element</span>
+      <span>Fake element</span>
+      <span>Fake element</span>
+      <span>Fake element</span>
+      <span>Fake element</span>
+      <span>Fake element</span>
+    </div>
+  );
+}
+
+export default Menu;
