+++ ScoutSuite/frontend/src/variables.scss
@@ -15,9 +15,11 @@ $default-grey-subtle: #e5e4e4;
 $default-grey-light: #cccccc;
 $white: #ffffff;
 
-$link_color: #4990e2;
-$success_color: #3fbd50;
-$warning_color: #fccc04;
-$issue_color: #fc6464;
+$link-color: #4990e2;
+$success-color: #3fbd50;
+$warning-color: #fccc04;
+$issue-color: #fc6464;
 
-$border_radius: 5px;
+$border-radius: 5px;
+$nav-header-height: 5rem;
+$nav-menu-width: 20rem;
